Vale.{X64,PPC64LE}.QuickCodes: erasing ranges
Ranges are for tagging terms for error reporting, and the fact that
they extract is arguably shady: their implementation is the internal
`FStarC.Compiler.Range.Type.range`, which is included in fstar.lib like
the rest of the compiler. This will be gone soon with a new build and
leaner fstar.lib, and this would fail.

Regardless of that, it's anyway cleaner to erase these to make the
generated OCaml simpler.